+++ src/stage_handlers/ready.rs
@@ -5,18 +5,21 @@ use ratatui::crossterm::event::{KeyCode, KeyEvent};
 use crate::{
     errors::Context,
     game::Game,
-    stage_handlers::{FRAME_DURATION_IDLE, GameplayHandler, Stage, StageHandler},
+    stage_handlers::{FRAME_DURATION_IDLE, GameplayHandler, InstructionsHandler, Stage, StageHandler},
 };
 
 pub struct ReadyHandler;
 
 impl StageHandler for ReadyHandler {
     fn handle_key_pressed_event(&mut self, game: &mut Game, key_event: KeyEvent) -> Option<Stage> {
-        if key_event.code == KeyCode::Enter {
-            game.start().context("Failed to start the game").ok()?;
-            return Some(Stage::Gameplay(GameplayHandler::new()));
+        match key_event.code {
+            KeyCode::Enter => {
+                game.start().context("Failed to start the game").ok()?;
+                Some(Stage::Gameplay(GameplayHandler::new()))
+            }
+            KeyCode::F(1) => Some(Stage::Instructions(InstructionsHandler)),
+            _ => None,
         }
-        None
     }
 
     fn time_before_next_tick(&mut self, _game: &mut Game) -> Duration {